Add jar targets for ondeviceintelligence to NeuralNetworks APEX
This change adds specific framework jars to be used when flag is enabled
or disabled.

For framework/bootclasspath we conditionally use
framework-ondeviceintelligence jar or the -platform variant based on
the build flag.

The asssociated files are in intermediate folder in frameworks/base
/packages/NeuralNetworks to allow for keeping file commit history.
Eventually these will be moved to the packages/modules directory after
the build flag is fully rolled out.

This CL also replicates the proto file to avoid dependencies on the
complete framework-protos library.

Flag: build.release_ondevice_intelligence_module
Bug: 376427781
API-Coverage-Bug: 376427781
Change-Id: Iefe67471e7e9f2db47fd8850c84b28ea1d0eee27